[tests-only][full-ci] Check multiple response (#8687)
* return early if the spaces list is empty

* check responses of each space during batch action

* fix space id extraction

* uncomment code

---------

+++ tests/e2e/support/objects/app-admin-settings/spaces/actions.ts
@@ -2,9 +2,9 @@ import { Page } from 'playwright'
 import util from 'util'
 import { locatorUtils } from '../../../utils'
 
-const spaceTrSelector = 'tr'
+const spaceTrSelector = '.spaces-table tbody > tr'
 const actionConfirmButton = '.oc-modal-body-actions-confirm'
-const spaceIdSelector = `[data-item-id="%s"] .spaces-table-btn-action-dropdown`
+const contextMenuSelector = `[data-item-id="%s"] .spaces-table-btn-action-dropdown`
 const spaceCheckboxSelector = `[data-item-id="%s"]:not(.oc-table-highlighted) input[type=checkbox]`
 const contextMenuActionButton = `.oc-files-actions-%s-trigger`
 const inputFieldSelector =
@@ -26,7 +26,7 @@ const spaceMemberList =
 
 export const getDisplayedSpaces = async (page): Promise<string[]> => {
   const spaces = []
-  const result = page.locator(spaceTrSelector)
+  const result = await page.locator(spaceTrSelector)
 
   const count = await result.count()
   for (let i = 0; i < count; i++) {
@@ -36,17 +36,16 @@ export const getDisplayedSpaces = async (page): Promise<string[]> => {
   return spaces
 }
 
-const clickOnContextMenuActionButton = async (args: {
+const performAction = async (args: {
   page: Page
-  id: string
   action: string
-  isBatchActions: boolean
+  context: string
+  id?: string
 }): Promise<void> => {
-  const { page, id, action, isBatchActions } = args
-  let context = '.batch-actions'
-  if (!isBatchActions) {
-    context = '.context-menu'
-    await page.locator(util.format(spaceIdSelector, id)).click()
+  const { page, action, context, id } = args
+
+  if (id && context === 'context-menu') {
+    await page.locator(util.format(contextMenuSelector, id)).click()
   }
 
   let contextMenuActionButtonSelector = null
@@ -73,83 +72,75 @@ const clickOnContextMenuActionButton = async (args: {
       throw new Error(`${action} not implemented`)
   }
   await page.waitForSelector(contextMenuActionButtonSelector)
-  await page.locator(context).locator(contextMenuActionButtonSelector).click()
+  await page.locator(`.${context}`).locator(contextMenuActionButtonSelector).click()
 }
 
 export const changeSpaceQuota = async (args: {
   page: Page
-  id: string
+  spaceIds: string[]
   value: string
   context: string
 }): Promise<void> => {
-  const { page, value, id, context } = args
-  const action = 'edit-quota'
-  const isBatchActions = context === 'batch-actions'
-  await clickOnContextMenuActionButton({ page, id, action, isBatchActions })
+  const { page, value, spaceIds, context } = args
+  await performAction({ page, action: 'edit-quota', context, id: spaceIds[0] })
+
   const searchLocator = await page.locator(spacesQuotaSearchField)
   await searchLocator.fill(value)
   await page.waitForSelector(selectedQuotaValueField)
   await page.locator(util.format(quotaValueDropDown, `${value} GB`)).click()
-  await waitForSpaceResponse({
+  await confirmAction({
     page,
     method: 'PATCH',
     statusCode: 200,
-    isBatchActions,
-    id,
+    spaceIds,
     actionConfirm: true
   })
 }
 
 export const disableSpace = async (args: {
   page: Page
-  id: string
+  spaceIds: string[]
   context: string
 }): Promise<void> => {
-  const { page, id, context } = args
-  const isBatchActions = context === 'batch-actions'
-  await clickOnContextMenuActionButton({ page, id, action: 'disable', isBatchActions })
-  await waitForSpaceResponse({
+  const { page, spaceIds, context } = args
+  await performAction({ page, action: 'disable', context, id: spaceIds[0] })
+  await confirmAction({
     page,
     method: 'DELETE',
     statusCode: 204,
-    isBatchActions,
-    id,
+    spaceIds,
     actionConfirm: false
   })
 }
 
 export const enableSpace = async (args: {
   page: Page
-  id: string
+  spaceIds: string[]
   context: string
 }): Promise<void> => {
-  const { page, id, context } = args
-  const isBatchActions = context === 'batch-actions'
-  await clickOnContextMenuActionButton({ page, id, action: 'restore', isBatchActions })
-  await waitForSpaceResponse({
+  const { page, spaceIds, context } = args
+  await performAction({ page, action: 'restore', context, id: spaceIds[0] })
+  await confirmAction({
     page,
     method: 'PATCH',
     statusCode: 200,
-    isBatchActions,
-    id,
+    spaceIds,
     actionConfirm: false
   })
 }
 
 export const deleteSpace = async (args: {
   page: Page
-  id: string
+  spaceIds: string[]
   context: string
 }): Promise<void> => {
-  const { page, id, context } = args
-  const isBatchActions = context === 'batch-actions'
-  await clickOnContextMenuActionButton({ page, id, action: 'delete', isBatchActions })
-  await waitForSpaceResponse({
+  const { page, spaceIds, context } = args
+  await performAction({ page, action: 'delete', context, id: spaceIds[0] })
+  await confirmAction({
     page,
     method: 'DELETE',
     statusCode: 204,
-    isBatchActions,
-    id,
+    spaceIds,
     actionConfirm: false
   })
 }
@@ -170,16 +161,13 @@ export const renameSpace = async (args: {
   value: string
 }): Promise<void> => {
   const { page, id, value } = args
-  const action = 'rename'
-  const isBatchActions = false
-  await clickOnContextMenuActionButton({ page, id, action, isBatchActions })
+  await performAction({ page, action: 'rename', context: 'context-menu', id })
   await page.locator(inputFieldSelector).fill(value)
-  await waitForSpaceResponse({
+  await confirmAction({
     page,
     method: 'PATCH',
     statusCode: 200,
-    isBatchActions,
-    id,
+    spaceIds: [id],
     actionConfirm: true
   })
 }
@@ -190,47 +178,44 @@ export const changeSpaceSubtitle = async (args: {
   value: string
 }): Promise<void> => {
   const { page, id, value } = args
-  const isBatchActions = false
-  await clickOnContextMenuActionButton({
-    page,
-    id,
-    action: 'edit-description',
-    isBatchActions
-  })
+  await performAction({ page, action: 'edit-description', context: 'context-menu', id })
   await page.locator(inputFieldSelector).fill(value)
-  await waitForSpaceResponse({
+  await confirmAction({
     page,
     method: 'PATCH',
     statusCode: 200,
-    isBatchActions,
-    id,
+    spaceIds: [id],
     actionConfirm: true
   })
 }
 
-const waitForSpaceResponse = async (args: {
+const confirmAction = async (args: {
   page: Page
   method: string
   statusCode: number
-  isBatchActions: boolean
-  id: string
+  spaceIds: string[]
   actionConfirm: boolean
 }): Promise<void> => {
-  const { page, method, statusCode, isBatchActions, id, actionConfirm } = args
+  const { page, method, statusCode, spaceIds, actionConfirm } = args
   let confirmButton = modalConfirmBtn
   if (actionConfirm) {
     confirmButton = actionConfirmButton
   }
+
+  const checkResponses = []
+  for (const id of spaceIds) {
+    checkResponses.push(
+      page.waitForResponse(
+        (resp) =>
+          resp.url().endsWith(encodeURIComponent(id)) &&
+          resp.status() === statusCode &&
+          resp.request().method() === method
+      )
+    )
+  }
+
   await page.waitForSelector(confirmButton)
-  await Promise.all([
-    page.waitForResponse(
-      (resp) =>
-        (isBatchActions || resp.url().endsWith(encodeURIComponent(id))) &&
-        resp.status() === statusCode &&
-        resp.request().method() === method
-    ),
-    page.locator(confirmButton).click()
-  ])
+  await Promise.all([...checkResponses, page.locator(confirmButton).click()])
 }
 
 export const openSpaceAdminSidebarPanel = async (args: {
